+++ CHANGES.rst
@@ -4,6 +4,19 @@
 Release Notes
 =============
 
+
+2.0.2 (unreleased)
+==================
+
+- Fix a bug in ``resample.Drizzle`` due to which initially, before adding
+  the first image, ``Drizzle.output_img`` is not filled with ``NaN`` when
+  ``fillval`` is either ``INDEF``, ``NAN``, ``None`` *and* the ``Drizzle``
+  object was initialized with ``out_img=None``. [#170]
+
+- Fixes a crash when ``Drizzle`` is initialized with ``disable_ctx``
+  set to ``True``. [#180]
+
+
 2.0.1 (2025-01-28)
 ==================
 
@@ -23,7 +36,7 @@ Release Notes
 - Deprecated module ``util``. New software should not import from this
   module as it will be removed in a future release. [#134]
 
-- Bug fix: exposure time was undefined when in_units were not cps. [#134]
+- Bug fix: exposure time was undefined when ``in_units`` were not "cps". [#134]
 
 - BUG FIX: ``cdrizzle.tdriz`` signature was out of date. [#134]